requirements.txt
pyserial
requests
flask
